requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
easyocr==1.7.2
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
flatbuffers==25.2.10
fonttools==4.58.2
fsspec==2025.5.1
humanfriendly==10.0
idna==3.10
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
lap==0.5.12
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.5
ninja==1.11.1.4
numpy==2.3.0
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
onnx==1.17.0
onnxruntime==1.22.0
onnxslim==0.1.57
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==25.0
pandas==2.3.0
pillow==11.2.1
protobuf==6.31.1
psutil==7.0.0
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pyparsing==3.2.3
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pytubefix==9.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
scikit-image==0.25.2
scipy==1.15.3
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
sympy==1.14.0
tifffile==2025.6.11
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
triton==3.3.1
typing_extensions==4.14.0
tzdata==2025.2
ultralytics==8.3.152
ultralytics-thop==2.0.14
urllib3==2.4.0
Werkzeug==3.1.3
